# openapi.yml
openapi: 3.0.1
info:
  title: mitum-currency - digest API
  description: >-
    This is API description of digest service of mitum-currency.
  contact:
    name: Spike^ekipS
    url: https://github.com/spikeekips/mitum-currency
    email: spikeekips@gmail.com
  license:
    name: GPL-3.0
    url: https://github.com/spikeekips/mitum-currency/blob/master/LICENSE
  version: 0.0.1
...
servers:
- url: https://localhost:54322
...
paths:
...
  /currency/{currency_id}:
    get:
      tags:
      - currency
      summary: Currency information
      operationId: currency
      parameters:
        - name: currency_id
          in: path
          description: currency unique id(or name)
          required: true
          schema:
            $ref: '#/components/schemas/CurrencyID'
      responses:
        500:
          description: problems in processing.
          content:
            application/problem+json:
              schema:
                $ref: '#/components/schemas/Problem'
        200:
          description: hal document of currency of *currency_id*
          content:
            application/hal+json:
              schema:
                $ref: '#/components/schemas/CurrencyHAL'
          headers:
            X-Rate-Limit:
              description: calls per hour allowed by the user
              schema:
                type: integer
                format: int32
            X-Rate-Remaining:
              description: remains request count
              schema:
                type: integer
                format: int32
            X-Rate-Reset:
              description: timestamp to reset limit
              schema:
                type: integer
                format: int64
components:
  schemas:
    Hint:
      type: string
      example: a015:0.0.1
    Problem:
      type: object
      required:
      - _hint
      - title
      - type
      - detail
      properties:
        _hint:
          allOf:
            - $ref: '#/components/schemas/Hint'
            - type: string
              example: a017:0.0.1
              default: a017:0.0.1
        title:
          type: string
          example: "sign"
        type:
          type: string
          example: "https://github.com/spikeekips/mitum-currency/problems/others"
        detail:
          type: string
          example: "invalid privatekey"
...
    HAL:
      type: object
      required:
      - _hint
      - _links
      properties:
        _hint:
          allOf:
            - $ref: '#/components/schemas/Hint'
            - type: string
              example: a016:0.0.1
              default: a016:0.0.1
        _embedded:
          type: object
          default: {}
        _links:
          type: object
          properties:
            self:
              allOf:
                - $ref: '#/components/schemas/HALLink'
                - type: object
                  properties:
                    href:
                      type: string
                      default: /
                      example: /
                    templated:
                      type: boolean
                      default: false
                      example: false
    HALLink:
      type: object
      required:
        - href
      properties:
        href:
          type: string
        templated:
          type: boolean
          description: The link is is templated link, not used without it's arguments.
          default: false
          example: false
...
    CurrencyHAL:
      allOf:
        - $ref: '#/components/schemas/HAL'
        - type: object
          properties:
             hint:
              type: object
              properties:
                name:
                  type: string
                  default: mitum-currency-currency-design
                  example: mitum-currency-currency-design
                hint:
                  type: string
                  default: a030:0.0.1
                  example: a030:0.0.1
             _embedded:
                $ref: '#/components/schemas/CurrencyDesign'
             _links:
                type: object
                properties:
                  self:
                    allOf:
                      - $ref: '#/components/schemas/HALLink'
                      - type: object
                        properties:
                          href:
                            type: string
                            default: /currency/XXX
                            example: /currency/XXX
    CurrencyID:
      description: currency unique id(or name)
      type: string
      example: XXX
...
    CurrencyDesign:
      type: object
      required:
      - _hint
      - amount
      - genesis_account
      - policy
      properties:
        _hint:
          allOf:
            - $ref: '#/components/schemas/Hint'
            - type: string
              default: a030:0.0.1
              example: a030:0.0.1
        amount:
          $ref: '#/components/schemas/Amount'
        genesis_account:
          allOf:
            - $ref: '#/components/schemas/AccountAddress'
            - description: genesis account address, which will hold genesis balance
        policy:
          $ref: '#/components/schemas/CurrencyPolicy'
    Amount:
      type: object
      required:
      - _hint
      - hash
      - amount
      - currency
      properties:
        _hint:
          allOf:
            - $ref: '#/components/schemas/Hint'
            - type: string
              default: a022:0.0.1
              example: a022:0.0.1
        hash:
          type: string
          format: hash
          example: yaJGkXgizfaB3HEAf9LWdWBC3Sh954pzuc2tNTM9PHG
        amount:
          type: string
          description: amount
          example: 33
          default: 33
        currency:
          $ref: '#/components/schemas/CurrencyID'
    CurrencyPolicy:
      description: currency policy
      type: object
      required:
      - _hint
      - new_account_min_balance
      - feeer
      properties:
        _hint:
          allOf:
            - $ref: '#/components/schemas/Hint'
            - type: string
              default: a036:0.0.1
              example: a036:0.0.1
        _hash:
          type: string
          format: hash
          example: yaJGkXgizfaB3HEAf9LWdWBC3Sh954pzuc2tNTM9PHG
        new_account_min_balance:
          allOf:
            - $ref: '#/components/schemas/Amount'
            - description: minimum balance for new account
        feeer:
          description: fee policy
          type: object
          oneOf:
            - $ref: '#/components/schemas/NilFeeer'
            - $ref: '#/components/schemas/FixedFeeer'
            - $ref: '#/components/schemas/RatioFeeer'
    NilFeeer:
      description: fee policy, which does not charge fee
      type: object
      required:
      - _hint
      properties:
        _hint:
          allOf:
            - $ref: '#/components/schemas/Hint'
            - type: string
              default: a031:0.0.1
              example: a031:0.0.1
        type:
          type: string
          example: 'nil'
          default: 'nil'
    FixedFeeer:
      description: fee policy, which does charge fixed amount
      type: object
      required:
      - _hint
      properties:
        _hint:
          allOf:
            - $ref: '#/components/schemas/Hint'
            - type: string
              default: a032:0.0.1
              example: a032:0.0.1
        type:
          type: string
          example: 'fixed'
          default: 'fixed'
        receiver:
          allOf:
            - $ref: '#/components/schemas/AccountAddress'
            - description: accound address for receving collected fee
        amount:
          allOf:
            - $ref: '#/components/schemas/Amount'
            - description: fee amount
    RatioFeeer:
      description: fee policy, which does charge fee by transfer amount
      type: object
      required:
      - _hint
      properties:
        _hint:
          allOf:
            - $ref: '#/components/schemas/Hint'
            - type: string
              default: a033:0.0.1
              example: a033:0.0.1
        type:
          type: string
          example: 'ratio'
          default: 'ratio'
        receiver:
          allOf:
            - $ref: '#/components/schemas/AccountAddress'
            - description: accound address for receving collected fee
        min:
          allOf:
            - $ref: '#/components/schemas/Amount'
            - description: minimum amounf of fee
        max:
          allOf:
            - $ref: '#/components/schemas/Amount'
            - description: maximum amounf of fee
...
    AccountAddress:
      description: account address
      type: string
      example: 'mc-node-010a:0.0.1'
...